factory & sim factory rename
